[DOP-10395] - move detect_hwm_store to separate module
